requirements.txt
absl-py==0.7.1
astor==0.8.0
certifi==2019.6.16
gast==0.2.2
grpcio==1.22.0
h5py==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
netron==3.3.0
numpy==1.17.0
protobuf==3.9.0
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
Werkzeug==0.15.5
